Rename local state file
