+++ src/material/mod.rs
@@ -4,13 +4,13 @@ mod conductivity;
 mod linear_elastic;
 mod loading_path;
 mod local_state;
-mod local_state_porous;
+mod local_state_porous_sld_liq;
 mod stress_strain;
 mod von_mises;
 pub use crate::material::conductivity::*;
 pub use crate::material::linear_elastic::*;
 pub use crate::material::loading_path::*;
 pub use crate::material::local_state::*;
-pub use crate::material::local_state_porous::*;
+pub use crate::material::local_state_porous_sld_liq::*;
 pub use crate::material::stress_strain::*;
 pub use crate::material::von_mises::*;
